able to create aggregation CSV

+++ simultaneous-grader/src/main/java/model/resultReport/ProblemResultDetails.java
@@ -2,7 +2,7 @@ package model.resultReport;
 
 import model.component.Student;
 
-public record ProblemResultDetails(String name, Student student, float passedPercent) {
+public record ProblemResultDetails(String name, Student student, double passedPercent) {
 	public String toCSVRow() {
 		return null;
 //		return ReportUtils.convertToCsvRow(student.idNumber(), student.fullName(), earnedPoints, passedFailed(), feedback);
